requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
billiard==4.2.2
Brotli==1.1.0
cachetools==5.5.2
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
fastapi==0.104.1
fonttools==4.60.1
frozenlist==1.8.0
google-api-core==2.27.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.71.0
greenlet==2.0.2
h11==0.16.0
html5lib==1.1
httplib2==0.31.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.2
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.9.0
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.3
redis==5.0.1
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.21
starlette==0.27.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.8.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
weasyprint==59.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.1
yarl==1.22.0
zopfli==0.2.3.post1
